Pass new commit to deploy action when sync commits
Without this, the deploy action will run with the same commit ID that started the sync action.
That results in deploying the site without the sync'd changes.
This attempts to avoid deploying when there is no new commit, and it passes the new commit to the deploy action when there is one.