Revert back to unofficial gh-pages action
The official action will not allow deploying a commit that is different from the commit that started the initial action.
In the case of the sync action, that is a problem when there is an updated file.
We will push a commit and want the newly pushed commit to be deployed.