Autogen category links for Reddit nesting (#14)
Customizes the result of the autogenerated sidebar to organize it the same way Reddit's wiki is organized.
This means linking the category created by folders to the "index" page for that folder located one level up with the same name as the folder. Also, removes the no longer needed sidebar item to the "index" page because it would be redundant (see method `updateAutogeneratedSiderbarForRedditWikiStructure`).