Update relative links to anchor from /
These links were broken if the page is located at a URL under a sub-path.
This applies the fix from gh-20 prior to merging that fix.
If the fix were merged before this change, there would be one commit per file and it would look like content changed in Reddit, attributed to an author, when it in fact didn't.

See gh-20